feat: 캐러셀 & 토스트 적용 및 테스트 페이지 제거
캐러셀과 토스트를 각각 상품 목록 페이지, 상품 상세 페이지에 구현함
더이상 불필요해진 테스트 페이지를 제거함
