+++ src/components/atoms/Toast.js
@@ -1,0 +1,14 @@
+import "components/atoms/Toast.css";
+
+import { Link } from "react-router-dom";
+
+export default function Toast() {
+  return (
+    <div className="z-10 fixed bottom-12 left-[50%] translate-x-[-50%] w-[48rem] px-8 py-5 rounded bg-stone-800 flex justify-between">
+      <span className="text-sm text-white">장바구니에 상품이 담겼습니다.</span>
+      <Link className="text-sm text-yellow-300" to="/cart">
+        바로가기
+      </Link>
+    </div>
+  );
+}
