--- src/styles/tests/Breadcrumb.css
@@ -1,14 +1,0 @@
-.breadcrumb-last-path {
-  font-weight: bold;
-}
-
-.breadcrumb-path {
-  font-weight: bold;
-  color: #707070;
-}
-
-.breadcrumb-a {
-  font-weight: bold;
-  text-decoration: none;
-  color: #707070;
-}
