--- src/styles/tests/Carousel.css
@@ -1,36 +1,0 @@
-.carousel-div {
-  position: relative;
-}
-
-.carousel-img {
-  width: 100%;
-}
-
-.carousel-span {
-  position: absolute;
-  top: 80%;
-  transform: translateX(50%);
-  right: 50%;
-  color: #ffffff90;
-}
-
-.carousel-button {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 3rem;
-  height: 3rem;
-  border: none;
-  border-radius: 50%;
-  background-color: #ffffff40;
-  font-size: 2rem;
-  color: #ffffff;
-}
-
-.carousel-button-left {
-  left: 5%;
-}
-
-.carousel-button-right {
-  right: 5%;
-}
